refactor: new permission components, useMemo
